Adds noise factory with tests. The factory is able to create Perlin, Billow and Ridged Noise.

+++ Masterthesis/Assets/Scripts/Random/Seed.cs
@@ -1,7 +1,3 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
 
 namespace SBaier.Master
 {
@@ -14,6 +10,6 @@ namespace SBaier.Master
 		}
 
 		public System.Random Random { get; }
-		public double SeedNumber { get; }
+		public int SeedNumber { get; }
 	}
 }
